+++ examples/slices.rs
@@ -14,22 +14,24 @@ fn main() {
 fn setup(mut cmd: Commands, server: Res<AssetServer>) {
     cmd.spawn((Camera2d, Transform::default().with_scale(Vec3::splat(0.1))));
 
-    cmd.spawn(AsepriteSliceBundle {
-        slice: "ghost_red".into(),
-        aseprite: server.load("ghost_slices.aseprite").into(),
-        transform: Transform::from_translation(Vec3::new(0., 0., 0.))
+    cmd.spawn((
+        AseSpriteSlice {
+            name: "ghost_red".into(),
+            aseprite: server.load("ball.aseprite"),
+        },
+        Transform::from_translation(Vec3::new(0., 0., 0.))
             .with_rotation(Quat::from_rotation_z(0.2)),
-        ..Default::default()
-    });
+    ));
 
-    cmd.spawn(AsepriteSliceBundle {
-        slice: "ghost_blue".into(),
-        sprite: Sprite {
+    cmd.spawn((
+        AseSpriteSlice {
+            name: "ghost_blue".into(),
+            aseprite: server.load("ball.aseprite"),
+        },
+        Sprite {
             flip_x: true,
             ..default()
         },
-        aseprite: server.load("ghost_slices.aseprite").into(),
-        transform: Transform::from_translation(Vec3::new(32., 0., 0.)),
-        ..Default::default()
-    });
+        Transform::from_translation(Vec3::new(32., 0., 0.)),
+    ));
 }